[CORDA-1970]: Refactored DI annotations.
